webapp/project: refactor start/restart/stop buttons, introduce a hook for the project state (and use some existing type info) and rename/reorg project-info and also use the restart button
