+++ src/smc-webapp/project/info/index.ts
