+++ src/smc-webapp/project/info/types.ts
