+++ src/smc-webapp/project/info/utils.ts
