+++ src/smc-webapp/project/page/content.tsx
@@ -39,7 +39,7 @@ import { DeletedFile } from "../deleted-file";
 import { KioskModeBanner } from "../../app/kiosk-mode-banner";
 import { Explorer } from "../explorer";
 import { ProjectNew } from "../new";
-import { ProjectInfoFC } from "../project-info";
+import { ProjectInfoFC } from "../info";
 import { ProjectLog } from "../history";
 import { ProjectSearch } from "../search/search";
 import { ProjectSettings } from "../settings";
